FEAT: add custom tabs for my wallet screen
