+++ src/routes.js
@@ -2,9 +2,8 @@ import React from 'react';
 import { View, Text } from 'react-native';
 import { Entypo } from '@expo/vector-icons';
 import { StackNavigator, SwitchNavigator, TabNavigator } from 'react-navigation';
-import { defineMessages } from 'react-intl';
+import { FormattedMessage } from 'react-intl';
 
-import { TopBar } from 'coupon-components-native';
 import LoginScreen from './containers/LogIn/LoginScreen';
 import StartScreen from './containers/Welcome/StartScreen';
 import RegisterScreen from './containers/Register/RegisterScreen';
@@ -12,13 +11,14 @@ import HomeScreen from './containers/Home/HomeScreen';
 import AuthLoadingScreen from './containers/AuthLoadingScreen/AuthLoadingScreen';
 import ProfileScene from './containers/Profile/ProfileScene';
 import CouponDetailScene from './containers/CouponDetail/CouponDetailScene';
-
-// Assets
-import arrow_left_c from './assets/images/arrow-left-c.png'
 import WalletScene from './containers/Wallet/WalletScene';
 import ExploreScene from './containers/Explore/ExploreScene';
 import NotificationsScene from './containers/Notifications/NotificationsScene';
 
+// Assets
+import arrow_left_c from './assets/images/arrow-left-c.png'
+
+// Auth Stack
 const AuthStack = StackNavigator({
   Welcome: { screen: StartScreen },
   Login: { screen: LoginScreen },
@@ -40,15 +40,16 @@ const AuthStack = StackNavigator({
   })
 });
 
+// Home Stack
 const HomeStack = StackNavigator({
   Home: { screen: HomeScreen },
   Profile: { screen: ProfileScene },
 },{
   navigationOptions: {
     header: null,
-    title: 'Today'
   }
 })
+
 
 const WalletStack = StackNavigator({
   Wallet: { screen: WalletScene },
@@ -60,8 +61,9 @@ const WalletStack = StackNavigator({
   }
 })
 
+// App Stack
 const AppStack = TabNavigator({
-  Home: { screen: HomeStack },
+  Home: { screen: HomeStack, navigationOptions: { title: <FormattedMessage id="commons.titles.today" /> }  },
   Wallet: { screen: WalletStack },
   Explore: { screen: ExploreScene },
   Notifications: { screen: NotificationsScene },
@@ -96,6 +98,7 @@ const AppStack = TabNavigator({
   })
 });
 
+// Modal Stack
 const ModalStack = StackNavigator({
   CouponDetails: { screen: CouponDetailScene }
 },{
